// src/api/generated/model/taskStatusEnum.ts
// @ts-nocheck
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * Dev Blackbox API
 * OpenAPI spec version: 0.1.1
 */

export type TaskStatusEnum = typeof TaskStatusEnum[keyof typeof TaskStatusEnum];


export const TaskStatusEnum = {
  BACKLOG: 'BACKLOG',
  TODO: 'TODO',
  IN_PROGRESS: 'IN_PROGRESS',
  DONE: 'DONE',
  CANCELED: 'CANCELED',
} as const;